Broken up OpenALSoundProvider.cpp
